+++ Assets/Runtime/Scripts/Engine/Modules/Rpg/Character/CharacterItem.cs
@@ -8,7 +8,7 @@ namespace io.github.thisisnozaku.idle.framework.Engine.Modules.Rpg
 {
     public class CharacterItem : RpgCharacterModifier
     {
-        public CharacterItem(long id, IdleEngine engine, string description, string[] usedSlots, Dictionary<string, Tuple<string, string>> modifications, Dictionary<string, List<String>> events) : base(engine, id, description, modifications, events)
+        public CharacterItem(double id, IdleEngine engine, string description, string[] usedSlots, Dictionary<string, Tuple<string, string>> modifications, Dictionary<string, List<String>> events) : base(engine, id, description, modifications, events)
         {
             this.UsedSlots = usedSlots;
         }
@@ -23,7 +23,7 @@ namespace io.github.thisisnozaku.idle.framework.Engine.Modules.Rpg
             private List<string> usedSlots = new List<string>();
             private Dictionary<string, List<string>> events = new Dictionary<string, List<string>>();
             private string description;
-            public override CharacterItem Build(IdleEngine engine, long id)
+            public override CharacterItem Build(IdleEngine engine, double id)
             {
                 return new CharacterItem(id, engine, description, usedSlots.ToArray(), modifications, events);
             }
